float을 bigDecimal 로 변경

+++ app/src/main/java/com/depromeet/muyaho/ui/addstock/AddStockInputFragment.kt
@@ -11,6 +11,7 @@ import com.depromeet.muyaho.databinding.FragmentAddStockInputBinding
 import com.depromeet.muyaho.util.NumberFormatUtil
 import com.depromeet.muyaho.widget.PriceEditText
 import dagger.hilt.android.AndroidEntryPoint
+import java.math.BigDecimal
 
 @AndroidEntryPoint
 class AddStockInputFragment :
@@ -100,24 +101,24 @@ class AddStockInputFragment :
     fun updatePrice() {
         val averagePrice = binding.petAveragePrice.price.let {
             if (it.isBlank()) {
-                0
+                0.toBigDecimal()
             } else {
-                it.toInt()
+                it.toBigDecimal()
             }
         }
 
         val stockCount = binding.petQuantity.price.let {
             if (it.isBlank()) {
-                0.0f
+                0.toBigDecimal()
             } else {
-                it.toFloat()
+                it.toBigDecimal()
             }
         }
 
-        var price: Float = averagePrice * stockCount
+        var price: BigDecimal = averagePrice * stockCount
         val isDollar = (vm.stockType == StockType.Overseas.full_name)
         if (isDollar) {
-            price *= 1200
+            price *= 1200.toBigDecimal()
         }
 
         binding.tvStockPrice.text = NumberFormatUtil.numWithComma(price)
